fix the IO-pairs of cp and get the result

+++ input-output-pairs/correct/cp_a/cp_n.c
@@ -23,7 +23,7 @@ int main(int argc, char** argv)
   nb_read = fread(buffer, 1, 100, from);
   while(0 < nb_read)
   {
-    int nb_written = fwrite(buffer, 1, nb_read, to);
+    int nb_written = fwrite(buffer, 1, (uintptr_t) nb_read, to);
     nb_read = fread(buffer, 1, 100, from);
   }
   fclose(from);
